Unblock CI + fix /t/[symbol] empty score-bar bug
ROOT CAUSE: ruff lint has been failing on every push since 2026-05-05,
gating Vercel + Fly deploys. Live site has been serving a 66-hour-old
build — universe-2,500, brand sweep, copy fixes, and Benzinga work were
all queued behind this.

LINT FIXES (12 ruff errors → 0):
- benzinga_feed.py: timezone.utc → UTC alias (UP017 ×4, mine)
- alerts.py: combine nested-if into single conjunction (SIM102)
- fear_greed.py: drop redundant int() — round() already returns int (RUF046)
- news_feed.py: refactor split-import that ruff was generating into a
  module-style alias (`benzinga_feed as bz`) — cleaner than the
  fight-with-isort version
- trial_abuse.py: typing.Deque → collections.deque (UP035, UP006)
- universe.py: typing.List/Tuple → list/tuple (UP035, UP006 ×4)

PUBLIC TICKER PAGE BUG (/t/[symbol]):
The page was reading `data.sub_trend`, `data.sub_rs`, etc. as top-level
fields — but the /api/ticker/{symbol} endpoint nests them inside
`data.breakdown.<key>.value`. Result: every factor bar rendered at 0%
width with a "—" value. Bad first impression on social-share clicks
(the OG card preview links here).

Fixed by updating the TickerData type + factor mapping to read from
`breakdown.*.value` — same shape the auth'd /app/ticker/[symbol] page
already uses correctly.

Verified locally: `ruff check` passes, 19/19 smoke tests green,
`tsc --noEmit` clean.

+++ backend/app/services/fear_greed.py
@@ -120,7 +120,7 @@ def compute_fear_greed(
         + 0.20 * regime_score
         + 0.15 * spy_score
     )
-    score = int(round(_clamp(composite)))
+    score = round(_clamp(composite))
     label, color = label_for(score)
 
     return {
